Picture in Picture interacts poorly with AirPlay/HDMI https://bugs.webkit.org/show_bug.cgi?id=147061 <rdar://problem/19192076>
Reviewed by Tim Horton.

When we are actively playing to an external target, the
picture in picture button should be hidden.

Also, the availability of picture in picture is also
dependent on AirPlay, so that we don't auto-pip when
we're displaying on a TV.

* Modules/mediacontrols/mediaControlsiOS.js:
(ControllerIOS.prototype.configureInlineControls): Call the update function.
(ControllerIOS.prototype.updatePictureInPictureButton): Add or remove a hidden class.
* html/MediaElementSession.cpp:
(WebCore::MediaElementSession::allowsPictureInPicture): Check Airplay status.

Identifier: 165109@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186979 268f45cc-cd09-0410-ab3c-d52691b4dbfc