AX: Selection change as a result of focusing an element may cause Safari to crash https://bugs.webkit.org/show_bug.cgi?id=147052 <rdar://problem/21778212>
Patch by Nan Wang <n_wang@apple.com> on 2015-07-20
Reviewed by Chris Fleizach.

Source/WebCore:

When focusing an element, it may trigger a deferred layout that invalidates the render
element, which will cause axObjectCache() to be a nullptr, and lead to a crash. Fix that
by using a RefPtr to hold the object and also caching the axObjectCache().

Test: platform/mac/accessibility/focus-crash.html

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::setFocused):
(WebCore::AccessibilityRenderObject::setSelectedRows):

LayoutTests:

* platform/mac/accessibility/focus-crash-expected.txt: Added.
* platform/mac/accessibility/focus-crash.html: Added.

Identifier: 165177@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187053 268f45cc-cd09-0410-ab3c-d52691b4dbfc