REGRESSION(r184376): [SOUP] Multiple assertions when downloading files https://bugs.webkit.org/show_bug.cgi?id=147039
Reviewed by Darin Adler.

No new tests; this will be covered once we enable the network process for API tests.

* platform/network/soup/ResourceHandleSoup.cpp:
(WebCore::ResourceHandle::releaseForDownload): Call RefPtrBase::relaxAdoptionRequirement so
that we don't assert when storing the non-adopted ResourceHandle in a RefPtr. The ref will
be dropped in ResourceHandle::cleanupSoupOperation, which as the comment says should always
run. HOPEFULLY.
(WebCore::ResourceHandle::continueWillSendRequest): Don't assert that client() is nonnull,
because the code clearly expects and handles the case where it is null.
(WebCore::ResourceHandle::continueDidReceiveResponse): Ditto; note that here client() will
always be null for a download.


Identifier: 165184@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187101 268f45cc-cd09-0410-ab3c-d52691b4dbfc