Object allocation sinking phase is performing needless HashMap copies https://bugs.webkit.org/show_bug.cgi?id=147159
Reviewed by Geoffrey Garen.

The points-to analyzer in the object allocation sinking phase is
currently performing copies of its allocation and pointers tables in
several places. While this is not a huge problem since those tables are
usually small and we are in the FTL path anyway, we still shouldn't be
doing such useless copying.

This patch also removes the DFGInsertOSRHintsForUpdate files that are
no longer needed with the new object sinking phase and should have been
removed in r186795.

* CMakeLists.txt:
* JavaScriptCore.vcxproj/JavaScriptCore.vcxproj:
* JavaScriptCore.vcxproj/JavaScriptCore.vcxproj.filters:
* JavaScriptCore.xcodeproj/project.pbxproj:
* dfg/DFGInsertOSRHintsForUpdate.cpp: Removed.
(JSC::DFG::insertOSRHintsForUpdate): Deleted.
* dfg/DFGInsertOSRHintsForUpdate.h: Removed.
* dfg/DFGObjectAllocationSinkingPhase.cpp:

Identifier: 165198@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187118 268f45cc-cd09-0410-ab3c-d52691b4dbfc