Remove hardcoded "internal" from Buildbot.prototype._normalizeQueueInfo. https://bugs.webkit.org/show_bug.cgi?id=147043
Patch by Jason Marcell <jmarcell@apple.com> on 2015-07-22
Reviewed by Daniel Bates and Alexey Proskuryakov.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/Buildbot.js:
(Buildbot.prototype._normalizeQueueInfo):
The _normalizeQueueInfo method now calls the defaultBranches getter instead of using hardcoded values
in cases where the queue.branch property is not set and the queue is not a combined queue.
* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/BuildbotCombinedQueueView.js:
(BuildbotCombinedQueueView):
Ensuring uniformity of repository branches within a combined queue.
* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/WebKitBuildbot.js:
(WebKitBuildbot.prototype.get defaultBranches):
Added; provides a WebKit-specific implementation that sets the queue.branches dictionary to have a
single entry for "openSource" as the default set of branches. Note that we intentionally omitted a
base class implementation of the defaultBranches getter so as to cause a JavaScript TypeError when a
sub-queue of a combined queue does not specify property branch and the derived Buildbot class does not
implement defaultBranches so that a person can either update the definition of the sub-queue or
implement defaultBranches in the derived Buildbot class they are using.

Identifier: 165244@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187172 268f45cc-cd09-0410-ab3c-d52691b4dbfc