[EFL] The "Missing Plug-in" buttons are not showing up on some flash contents https://bugs.webkit.org/show_bug.cgi?id=147191
Patch by Sungmann Cho <sungmann.cho@navercorp.com> on 2015-07-22
Reviewed by Gyuyoung Kim.

Currently, WebKitEFL doesn't show the "Missing Plug-in" buttons if the plugin-related tags
don't have a "type" attribute. In such a case, WebCore tries to guess the MIME type from
the extensions by using MIMETypeRegistry::getMIMETypeForExtension(). For WebKitEFL,
MIMETypeRegistry::getMIMETypeForExtension() goes through |extensionMap|, which is a simple
array of <extension, mime type>, looking for the mime type for the given extension.
But |extensionMap| in MIMETypeRegistryEfl.cpp doesn't have the information for ".swf",
so WebCore fails to guess the MIME type and regard the content type as ObjectContentFrame,
not ObjectContentNetscapePlugin.

* platform/efl/MIMETypeRegistryEfl.cpp:

Identifier: 165269@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187206 268f45cc-cd09-0410-ab3c-d52691b4dbfc