Web Inspector: Timeline's Current Time does not jump forward to new start time when starting a new recording, causes timeline to appear delayed and broken https://bugs.webkit.org/show_bug.cgi?id=147204
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-07-22
Reviewed by Timothy Hatcher.

* UserInterface/Views/TimelineRecordingContentView.js:
(WebInspector.TimelineRecordingContentView.prototype._recordingTimesUpdated):
We were skipping a timeline, but it might have a new event record with
a new start time. This ensures we get an updated current time which makes
sense, and we jump forward to that time in the main timeline.

Identifier: 165272@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187209 268f45cc-cd09-0410-ab3c-d52691b4dbfc