JavaScriptCore bmalloc should not register its malloc zone more than once https://bugs.webkit.org/show_bug.cgi?id=147273
Reviewed by Andreas Kling.

This was a goof: The Zone constructor, by virtue of running automatically,
was registering a Zone inside the analysis process.

* bmalloc/Zone.cpp:
(bmalloc::remoteRead): Clarify that the pointer is remote.

(bmalloc::enumerator):
(bmalloc::Zone::Zone):
* bmalloc/Zone.h: Separate the normal constructor and the remote constructor.
The remote constructor skips zone registration since its goal is not
to register a zone in the current process or do any allocation but rather
to mirror the bytes of the zone from the target process.


Identifier: 165359@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187360 268f45cc-cd09-0410-ab3c-d52691b4dbfc