Bounds in InteractionInformationAtPosition should be always in main frame coordinate space. https://bugs.webkit.org/show_bug.cgi?id=147372
When we prepare the bounds for InteractionInformationAtPosition, we should convert the rect to
main frame space since WKContent in UIProcess expects it to be in the web view space.

Patch by Yongjun Zhang <yongjun_zhang@apple.com> on 2015-07-28
Reviewed by Tim Horton.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::getPositionInformation): Convert the bounding rect to main frame space if the element is inside a sub-frame.

Identifier: 165468@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187517 268f45cc-cd09-0410-ab3c-d52691b4dbfc