[GTK] Fix empty space in popup menus when first item is selected https://bugs.webkit.org/show_bug.cgi?id=147358
Reviewed by Martin Robinson.

It happens when one of the first elements are selected, because we
try to center the current selection on the combo element. When the
menu is large enough to make the element centered we need to leave
empty space at the beginning. This can be fixed by not scrolling
to center the selected item when it's above the combo
element. This ensure the selected item will always be visible,
even if it's not centered. If the selected item is already
centered or below the combo element, the behaviour doesn't change
and we scroll to center it.

* UIProcess/gtk/WebPopupMenuProxyGtk.cpp:
(WebKit::WebPopupMenuProxyGtk::showPopupMenu):

Identifier: 165491@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187542 268f45cc-cd09-0410-ab3c-d52691b4dbfc