shouldParseTelephoneNumbersInNode() does not need to check for Comment nodes https://bugs.webkit.org/show_bug.cgi?id=147630
Reviewed by Sam Weinig.

shouldParseTelephoneNumbersInNode() does not need to check for Comment
nodes. We already know the input is a ContainerNode and a Comment is
NOT a ContainerNode. Also, we were already doing a isElementNode()
check before calling disallowTelephoneNumberParsing().

I also updated the function to use a for loop for clarity and dropped
the isElementNode() check as calling hasTagName() / isLink() on a
ContainerNode is as fast all calling it on an Element nowadays.

* html/parser/HTMLTreeBuilder.cpp:
(WebCore::disallowTelephoneNumberParsing):
(WebCore::shouldParseTelephoneNumbersInNode):


Identifier: 165710@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187893 268f45cc-cd09-0410-ab3c-d52691b4dbfc