Subframes with no current HistoryItem should not prevent page-caching https://bugs.webkit.org/show_bug.cgi?id=147649 <rdar://problem/21614832>
Reviewed by Andreas Kling.

Source/WebCore:

Subframes with no current HistoryItem should not prevent page-caching.
We need one for the main frame as this is the key in the PageCache.
However, there is no reason to require one for subframes.

This is a common reason for page-caching failures nowadays.

Frames do no have a current HistoryItem until something has been loaded in them.

Test: http/tests/navigation/page-cache-iframe-no-current-historyItem.html

* history/PageCache.cpp:
(WebCore::logCanCacheFrameDecision):
(WebCore::PageCache::canCachePageContainingThisFrame):

LayoutTests:

Add a layout test that covers page-caching of a page that contains an
iframe with no current HistoryItem. The sub-frame has no current
HistoryItem because its URL is invalid: "http://".

* http/tests/navigation/page-cache-iframe-no-current-historyItem-expected.txt: Added.
* http/tests/navigation/page-cache-iframe-no-current-historyItem.html: Added.
New test.

* platform/mac-wk1/TestExpectations:
Skip the new test on Mac WK1. The page is still not cacheable on Mac WK1
because WebFrameLoaderClient::canCachePage() returns false if the frame
does not contain HTML at the moment.


Identifier: 165731@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@187921 268f45cc-cd09-0410-ab3c-d52691b4dbfc