[ES6] Class parser does not allow methods named set and get. https://bugs.webkit.org/show_bug.cgi?id=147150
Reviewed by Oliver Hunt.

Source/JavaScriptCore:

The bug was caused by parseClass assuming identifiers "get" and "set" could only appear
as the leading token for getter and setter methods. Fixed the bug by generalizing the code
so that we only treat them as such when it's followed by another token that could be a method name.

* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseClass):

LayoutTests:

Added a regression test and rebaselined a test.

* js/class-syntax-method-names-expected.txt: Added.
* js/class-syntax-method-names.html: Added.
* js/class-syntax-semicolon-expected.txt: Rebaselined as the error message got improved.
* js/script-tests/class-syntax-method-names.js: Added.
* js/script-tests/class-syntax-semicolon.js:


Identifier: 165798@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188018 268f45cc-cd09-0410-ab3c-d52691b4dbfc