CSSSegmentedFontFace::fontRanges() does not handle duplicate fonts correctly https://bugs.webkit.org/show_bug.cgi?id=147765
Reviewed by Filip Pizlo.

CSSSegmentedFontFace::fontRanges() was trying to hash on FontDescriptors by
picking a few specific pieces of data out of the FontDescriptor, computing
a hash on it, and using that unsigned as a key in a HashMap. This has two
problems: it doesn't handle equality correctly, as hash collisions cannot
depend on an equality operator to dedup, and it doesn't hash on all the
members of a FontDescription.

Instead, this HashMap should use FontDescriptionKey, which represents a
FontDescription, and is designed exactly for the purpose of being used as a
key in a HashMap.

No new tests because there is no behavior change (because a problem occurs
when two different FontDescriptions hash to the same value, which is rare).

* css/CSSSegmentedFontFace.cpp:
(WebCore::CSSSegmentedFontFace::fontRanges):
* css/CSSSegmentedFontFace.h:
* platform/graphics/FontCache.h:
(WebCore::FontDescriptionKeyHash::hash):
(WebCore::FontDescriptionKeyHash::equal):

Identifier: 165861@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188114 268f45cc-cd09-0410-ab3c-d52691b4dbfc