Fix GTK clean build after r187997 https://bugs.webkit.org/show_bug.cgi?id=147723
Patch by Alex Christensen <achristensen@webkit.org> on 2015-08-06
Reviewed by Gyuyoung Kim.

* CMakeLists.txt:
r187997 made it so WebCoreTestSupport does not link with WebCore, which is necessary to get DumpRenderTree to link.
It also made it so WebCoreTestSupport is no longer dependent on WebCore, which causes it to build at the same time.
This caused problems because WebCoreTestSupport uses headers that are generated for WebCore, such as WebKitFontFamilyNames.h.
Adding a dependency makes it so that WebCoreTestSupport waits until WebCore is done compiling before compiling,
which is what it used to do, but it does not cause linking problems with DumpRenderTree.

Identifier: 165863@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188116 268f45cc-cd09-0410-ab3c-d52691b4dbfc