[GTK] Test  /webkit2/WebKitWebView/submit-form is flaky https://bugs.webkit.org/show_bug.cgi?id=147727
Reviewed by Sergio Villar Senin.

I think it was not this test in particular, but
/webkit2/WebKitWebView/custom-charset that is affecting the
others. This is because changing the encoding reloads the page,
but we don't wait for the page to be reloaded, so when the test
finishes and other test starts the web process is still reloading
the page.

* Scripts/run-gtk-tests:
(TestRunner): Unskip /webkit2/WebKitWebView/submit-form.
* TestWebKitAPI/Tests/WebKit2Gtk/TestWebKitWebView.cpp:
(testWebViewCustomCharset): Wait until page is reloaded after
changing the charset.

Identifier: 165938@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188207 268f45cc-cd09-0410-ab3c-d52691b4dbfc