Occasional failure in v8-v6/v8-raytrace.js.ftl-eager https://bugs.webkit.org/show_bug.cgi?id=147165
Reviewed by Saam Barati.

The object allocation sinking phase was not properly checking that a
MultiGetByOffset was safe to lower before lowering it.
This makes it so that we only lower MultiGetByOffset if it only loads
from direct properties of the object, and considers it as an escape in
any other case (e.g. a load from the prototype).

It also ensure proper conversion of MultiGetByOffset into
CheckStructureImmediate when needed.

* dfg/DFGObjectAllocationSinkingPhase.cpp:
* ftl/FTLLowerDFGToLLVM.cpp:
(JSC::FTL::DFG::LowerDFGToLLVM::checkStructure):
    We were not compiling properly CheckStructure and
    CheckStructureImmediate nodes with an empty StructureSet.
* tests/stress/sink-multigetbyoffset.js: Regression test.


Identifier: 166172@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188507 268f45cc-cd09-0410-ab3c-d52691b4dbfc