Cannot change WKWebView scrollView deceleration rate https://bugs.webkit.org/show_bug.cgi?id=148086 <rdar://problem/21893685>
Reviewed by Simon Fraser.

Updates the preferred scroll deceleration factor when the scrollView's decelerationRate is explicitly set by the
user, so when snap scrolling is inactive and [WKWebView scrollViewWillBeginDragging] sets its scrollView's
decelerationRate to the preferred decelerationRate, it will be set to the updated value.

* UIProcess/ios/WKScrollView.mm:
(-[WKScrollView setDecelerationRate:]): Overrides setDecelerationRate to update _preferredScrollDecelerationFactor.

Identifier: 166203@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188541 268f45cc-cd09-0410-ab3c-d52691b4dbfc